Do not override globals.flutterVersion
